removed endpoint hashing, changed shelf structure, refactored methods into lib files
